refactor: Extract YOLOController from ImageAnnotator (Phase 4c)
Phase 4c completes Phase 4 in the modular refactoring: YOLO training
and prediction methods + the TrainingThread worker move into a
dedicated controller. With this PR all three inference engines
(SAM / DINO / YOLO) live in their own modules.

New controllers/yolo_controller.py (539 lines) — YOLOController(QObject)
plus the TrainingThread(QThread) class.

Methods relocated:
- Menu: setup_yolo_menu
- Training data + model loading: load_yolo_model, prepare_yolo_dataset,
  load_yolo_yaml, save_yolo_model, load_prediction_model,
  initialize_yolo_trainer
- Training execution: show_train_dialog, start_training,
  training_finished
- Prediction: set_confidence_threshold, show_predict_dialog,
  run_predictions, predict_single_image, process_yolo_results

State (yolo_trainer, training_thread, training_dialog) stays on the
main window for consistency. ImageAnnotator.__init__ still sets
yolo_trainer = None; setup_yolo_menu (delegated to the controller) is
called from __init__ after the menu bar exists.

Dead top-level imports removed from annotator_window.py:
- cv2 (only used by process_yolo_results)
- numpy (only used by process_yolo_results)
- QThread, pyqtSignal (only used by TrainingThread, now in YOLOController)
- QDoubleSpinBox (only used by show_predict_dialog)
- LoadPredictionModelDialog, TrainingInfoDialog, YOLOTrainer (referenced
  by methods that moved)

annotator_window.py: 3,194 → 2,757 lines (-437, -13.7%).
Total reduction vs original master: 5,900 → 2,757 (-53.3%).

More than half the original monolith now lives in focused controller
modules:
- controllers/io_controller.py        338
- controllers/project_controller.py   548
- controllers/image_controller.py     869
- controllers/sam_controller.py       267
- controllers/dino_controller.py      796
- controllers/yolo_controller.py      539
- core/image_utils.py                  74
- ui/theme.py                          68

Tests: all 94 pass.
